src/multas_aplicadas_acoes_fiscalizacao/raw/requirements.txt
requests
beautifulsoup4
pandas
google-cloud-storage
google-cloud-bigquery
pyarrow